# Fix vocab size mismatch in inference due to added pad token